test(assert): merge test classes
- merge both ProcessEngineTestsTest classes into one

related to CAM-9793
